move logging config out of Dispatcher
